requirements.txt
pytest
pydantic
pre-commit
mypy
flake8
fastlite
pyyaml
